Extract and rename method
