+++ app/src/ui/app-error.tsx
@@ -19,6 +19,7 @@ import { ErrorWithMetadata } from '../lib/error-with-metadata'
 import { RetryActionType, RetryAction } from '../models/retry-actions'
 import { Ref } from './lib/ref'
 import memoizeOne from 'memoize-one'
+import { parseCarriageReturn } from '../lib/parse-carriage-return'
 
 interface IAppErrorProps {
   /** The list of queued, app-wide, errors  */
@@ -51,7 +52,7 @@ interface IAppErrorState {
  */
 export class AppError extends React.Component<IAppErrorProps, IAppErrorState> {
   private dialogContent: HTMLDivElement | null = null
-  private formatGitErrorMessage = memoizeOne(carriageReturnFormatter)
+  private formatGitErrorMessage = memoizeOne(parseCarriageReturn)
 
   public constructor(props: IAppErrorProps) {
     super(props)
@@ -329,43 +330,4 @@ function isErrorWithMetaData(error: Error): error is ErrorWithMetadata {
 
 function isGitError(error: Error): error is GitError {
   return error instanceof GitError
-}
-
-function carriageReturnFormatter(text: string) {
-  const lines = new Array<string>('')
-  const crOrLf = /[\r\n]/gm
-
-  let lineIx = 0
-  let columnIx = 0
-  let p = 0
-
-  function merge(s: string) {
-    const line = lines[lineIx]
-    const before = line.substring(0, columnIx)
-    const after = line.substring(columnIx + s.length)
-    columnIx += s.length
-    lines[lineIx] = `${before}${s}${after}`
-  }
-
-  let m
-
-  while ((m = crOrLf.exec(text)) !== null) {
-    if (m.index > p) {
-      merge(text.substring(p, m.index))
-    }
-
-    if (m[0] === '\r') {
-      columnIx = 0
-    } else if (m[0] === '\n') {
-      lines[++lineIx] = ''
-    }
-
-    p = m.index + 1
-  }
-
-  if (p < text.length) {
-    merge(text.substring(p))
-  }
-
-  return lines.join('\n')
 }
